requirements.txt
einops==0.4.1
hydra-core==1.2.0
librosa==0.9.1
matplotlib==3.5.2
numpy==1.22.4
pandas==1.4.2
pytorch_lightning==1.6.4
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.9.0
torchvision==0.12.0
tqdm==4.64.0
wandb==0.12.17
x_transformers==0.26.6